// web/src/api/generated/models/post-detail-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PostDetailResponse {
    'id'?: number;
    'title'?: string;
    'content'?: string;
    'tags'?: Array<string>;
    'thumbNum'?: number;
    'favourNum'?: number;
    'commentNum'?: number;
    'userId'?: number;
    'ipAddress'?: string;
    'postType'?: string;
    'createTime'?: string;
    'updateTime'?: string;
    'hasThumb'?: boolean;
    'hasFavour'?: boolean;
}