// web/src/api/generated/models/comment-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CommentResponse {
    'id'?: number;
    'postId'?: number;
    'userId'?: number;
    'content'?: string;
    'ipAddress'?: string;
    'createTime'?: string;
    'updateTime'?: string;
}